+++ src/main/java/com/teamproject1/scuoledevelhope/types/dtos/BaseResponseElement.java
@@ -4,10 +4,6 @@ import org.springframework.http.HttpStatus;
 
 public class BaseResponseElement<T> extends BaseResponse {
     private T element;
-    private int page ;
-    private int pageSize ;
-    private long totalElements ;
-    private int totalPages ;
 
     public BaseResponseElement() {
 
@@ -24,37 +20,5 @@ public class BaseResponseElement<T> extends BaseResponse {
     }
     public T getElement() {
         return element;
-    }
-
-    public int getPage() {
-        return page;
-    }
-
-    public void setPage(int page) {
-        this.page = page;
-    }
-
-    public int getPageSize() {
-        return pageSize;
-    }
-
-    public void setPageSize(int pageSize) {
-        this.pageSize = pageSize;
-    }
-
-    public long getTotalElements() {
-        return totalElements;
-    }
-
-    public void setTotalElements(long totalElements) {
-        this.totalElements = totalElements;
-    }
-
-    public int getTotalPages() {
-        return totalPages;
-    }
-
-    public void setTotalPages(int totalPages) {
-        this.totalPages = totalPages;
     }
 }
