+++ src/main/java/com/teamproject1/scuoledevelhope/types/dtos/BaseResponseList.java
@@ -4,21 +4,16 @@ import org.springframework.http.HttpStatus;
 
 import java.util.List;
 
-public class BaseResponseList<T> extends BaseResponse {
+public class BaseResponseList<T> extends Pagination {
 
     List<T> elements;
-
-    private int page ;
-    private int pageSize ;
-    private long totalElements ;
-    private int totalPages ;
 
     public BaseResponseList() {
 
     }
 
-    public BaseResponseList(HttpStatus httpStatus, String message, String description, List<T> elements) {
-        super(httpStatus, message, description);
+    public BaseResponseList(HttpStatus httpStatus, String message, String description, int page, int pageSize, long totalElements, int totalPages, List<T> elements) {
+        super(httpStatus, message, description, page, pageSize, totalElements, totalPages);
         this.elements = elements;
     }
 
@@ -28,37 +23,5 @@ public class BaseResponseList<T> extends BaseResponse {
 
     public List<T> getElements() {
         return elements;
-    }
-
-    public int getPage() {
-        return page;
-    }
-
-    public void setPage(int page) {
-        this.page = page;
-    }
-
-    public int getPageSize() {
-        return pageSize;
-    }
-
-    public void setPageSize(int pageSize) {
-        this.pageSize = pageSize;
-    }
-
-    public long getTotalElements() {
-        return totalElements;
-    }
-
-    public void setTotalElements(long totalElements) {
-        this.totalElements = totalElements;
-    }
-
-    public int getTotalPages() {
-        return totalPages;
-    }
-
-    public void setTotalPages(int totalPages) {
-        this.totalPages = totalPages;
     }
 }
